Carry native pointers through the ARM64 graphics path

+++ include/libforest/gbi_extensions.h
@@ -21,8 +21,8 @@ extern "C" {
 
 #ifndef _GBI_RUNTIME_PTR_HELPERS
 #define _GBI_RUNTIME_PTR_HELPERS
-unsigned int pc_gbi_pack_runtime_ptr(uintptr_t addr, int is_ptr, const char* expr, const char* file, int line);
-uintptr_t pc_gbi_unpack_runtime_ptr(unsigned int packed);
+uintptr_t pc_gbi_pack_runtime_ptr(uintptr_t addr, int is_ptr, const char* expr, const char* file, int line);
+uintptr_t pc_gbi_unpack_runtime_ptr(uintptr_t packed);
 #endif
 
 #ifdef PC_64BIT
